+++ crates/bitwarden-policies/src/lib.rs
@@ -5,8 +5,7 @@ uniffi::setup_scaffolding!();
 #[cfg(feature = "uniffi")]
 mod uniffi_support;
 
-pub mod enforced_policy;
-pub mod filter;
+pub mod enforcement;
 mod master_password_policy_response;
 mod models;
 mod policy_client;
@@ -14,8 +13,9 @@ pub mod policy_overrides;
 mod policy_type;
 mod registry;
 
-pub use enforced_policy::{EnforcedCombinedPolicy, EnforcedPolicy};
-pub use filter::{NoData, Policy, PolicyAggregate, PolicyData};
+pub use enforcement::{
+    EnforcedCombinedPolicy, EnforcedPolicy, NoData, Policy, PolicyAggregate, PolicyData,
+};
 pub use master_password_policy_response::MasterPasswordPolicyResponse;
 pub use models::{OrganizationUserPolicyContext, PolicyView};
 pub use policy_client::{PoliciesClientExt, PolicyClient};
